Bug 835692: Convert TransportLayerDtls::kMaxDigestLength to be a size_t, since that's how it's used & what it's compared to. r=ekr
UltraBlame original commit: bf9886e0d35620c31ed2586a4a2fc8c316ddaa5e

+++ media/mtransport/transportlayerdtls.h
@@ -59,7 +59,7 @@ class TransportLayerDtls : public TransportLayer {
 
   enum Role { CLIENT, SERVER};
   enum Verification { VERIFY_UNSET, VERIFY_ALLOW_ALL, VERIFY_DIGEST};
-  const static int kMaxDigestLength = HASH_LENGTH_MAX;
+  const static size_t kMaxDigestLength = HASH_LENGTH_MAX;
 
   
   void SetRole(Role role) { role_ = role;}
